add run/action tests, update tests now that no args are allowed, remove old file
